Change Mass struct to use double for value to improve precision

+++ stella_invicta_headless/StellaInvicta/Components/Mass.cs
@@ -6,7 +6,7 @@ namespace StellaInvicta.Components;
 /// </summary>
 /// <param name="Value">The mass value in kilograms. Defaults to 0 if not specified.</param>
 /// <remarks>
-/// This is a record struct that holds a single float value representing mass.
+/// This is a record struct that holds a single double value representing mass.
 /// The struct is immutable and implements value-based equality comparisons.
 /// </remarks>
-public record struct Mass(float Value = 0);
+public record struct Mass(double Value = 0);
